\name{MakeFig}
\alias{MakeFig}
\title{fig positions for graphics}
\description{Calculates positions for figure that consist of multiple panels}
\usage{MakeFig(nfig, border = c(0, 1, 0, 1), nrow = NULL, ncol = NULL)}
\arguments{
  \item{nfig}{number of figures}
  \item{border}{relative graphic borders in which the figures should be placed}
  \item{nrow}{number of rows to arrange the figures}
  \item{ncol}{number of cols to arrange the figures}
}

\value{A list with positions for each figure and number of rows and columns}

\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}





