\name{SOfireA-package}
\alias{SOfireA-package}
\alias{SOfireA}
\docType{package}
\title{Satellite Observations for Fire Activity}
\description{SOFIA (Satellite Observations for Fire Activity) is an empirical modelling concept to predict wildfire burned area based on satellite and climate data. The package implements the basic SOFIA model structure, and functions to optimize, and plot SOFIA models.}

\details{
\tabular{ll}{
Package: \tab SOfireA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-11-18\cr
License: \tab GPL-2\cr
}

}

\author{
Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]

Maintainer: Matthias Forkel <matthias.forkel@geo.tuwien.ac.at>
}

\references{
Forkel, M., et al. (2016): in prep.
}

\keyword{ package }
\seealso{
http://greenbrown.r-forge.r-project.org/
}

